+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](semver.md).
 ### Added
 * Link Value to RksvBill
 * `alphanumericalRegex` global function
+* `replaceAllOf` extension for String
 ### Changed
 ### Deprecated
 * AustrianBill, use RksvBill instead